queue_app/requirements.txt
Flask
requests
pandas
numpy
pyyaml
# tornado==4.5.3